Note-Worthy App

+++ Unit-7 (Backend)/Unit7 Coding Evaluation/note_worthy/Backend/handlers/noteData.js
@@ -8,9 +8,10 @@ async function GetNotes(req,res){
 
     var token = req.headers.token;
     
-    const data = await noteData.find().populate("user")
-
+    
     if(token){
+        const decoded_user = jwt.verify(token, SECRET);
+        const data = await noteData.find({user:decoded_user.id}).populate("user")
         return res.send(data)
     }else{
         return res.send("User Not Logged In")
@@ -23,6 +24,8 @@ async function Postnotes(req,res){
 
 
     if(token){
+        const decoded_user = jwt.verify(token, SECRET);
+        note.user = decoded_user.id
         data = await noteData.create(note)
         return res.send({"data":data})
     }else{
@@ -41,6 +44,9 @@ async function UpdateNote(req,res){
     let {note: note_data} = req.body
 
     if(token){
+    const decoded_user = jwt.verify(token, SECRET);
+    note_data.user = decoded_user.id
+        
     let note = await noteData.findById(id)
 
     for (const [key, value] of Object.entries(note_data)) {
@@ -49,8 +55,6 @@ async function UpdateNote(req,res){
 
     await note.save()
     return res.send("Post Updated")
-    }else{
-        return res.send("User Not Logged In")
     }
 }
 
@@ -61,8 +65,15 @@ async function DeleteNote(req,res){
     let { id }  = req.params;
 
     if(token){
+        const decoded_user = jwt.verify(token, SECRET);
+        let note = await noteData.findById(id)
+
+        if(note.user = decoded_user.id){
         await noteData.findByIdAndDelete(id)
         return res.send("Post Deleted")
+        }else{
+            return res.send("You Can not Delete Other's Post")
+        }
     }else{
         return res.send("User Not Logged In")
     }
